Rename, add test
